apollot broken

+++ server/schemas/resolvers.js
@@ -9,7 +9,6 @@ const { ObjectId } = require("mongodb");
 require("dotenv").config();
 const bcrypt = require("bcryptjs");
 
-
 cloudinary.config({
   cloud_name: process.env.CLOUDINARY_NAME,
   api_key: process.env.CLOUDINARY_API_KEY,
@@ -34,7 +33,7 @@ const resolvers = {
       }
       throw new AuthenticationError("Authentication error");
     },
-    
+
     me: async (parent, args, context) => {
       if (context.user) {
         return await User.findOne({ _id: context.user._id });
@@ -61,7 +60,7 @@ const resolvers = {
         return null; // Post not found
       }
 
-      console.log("post found", post);
+      console.log("post found");
 
       return post;
     },
@@ -91,36 +90,23 @@ const resolvers = {
 
     // Add a post to a capsule by a logged in user
     uploadPost: async (parent, { file }) => {
-        try {
-          console.log("uploading post...", file)
-          // Upload the image to Cloudinary
-          const uploadResult = await cloudinary.uploader.upload(file, {
-            folder: 'wedloc', // Specify the folder in Cloudinary
-          });
-  
-          console.log("uploadResult", uploadResult)
-          // Return the result of the Cloudinary upload
-          return {
-            public_id: uploadResult.public_id,
-            secure_url: uploadResult.secure_url,
-          };
-        } catch (error) {
-          throw new Error('Error uploading image: ' + error.message);
-        }
-      },
+      try {
+        console.log("uploading post...", file);
+        // Upload the image to Cloudinary
+        const uploadResult = await cloudinary.uploader.upload(file, {
+          folder: "wedloc", // Specify the folder in Cloudinary
+        });
 
-
-    //   await Capsule.findOneAndUpdate(
-    //     { _id: capsuleId },
-    //     { $addToSet: { posts: post._id } }
-    //   );
-    //   await User.findOneAndUpdate(
-    //     { _id: context.user._id },
-    //     { $addToSet: { posts: post._id } }
-    //   );
-    //   return post;
-    // },
-    // Delete a post by a logged in user
+        console.log("uploadResult", uploadResult);
+        // Return the result of the Cloudinary upload
+        return {
+          public_id: uploadResult.public_id,
+          secure_url: uploadResult.secure_url,
+        };
+      } catch (error) {
+        throw new Error("Error uploading image: " + error.message);
+      }
+    },
     deletePost: async (parent, { postId }, context) => {
       if (context.user) {
         const post = await Post.findOneAndDelete({
@@ -152,13 +138,17 @@ const resolvers = {
       return newChat;
     },
 
-    addUser: async (parent, { username, email, password, firstName, lastName }) => {
+    addUser: async (
+      parent,
+      { username, email, password, firstName, lastName }
+    ) => {
       try {
-        const user = await User.create({ username, 
-          email, 
-          password, 
-          firstName, 
-          lastName 
+        const user = await User.create({
+          username,
+          email,
+          password,
+          firstName,
+          lastName,
         });
         const token = signToken(user);
         return { token, user };
@@ -168,7 +158,11 @@ const resolvers = {
       }
     },
 
-    updateUser: async (parent, { firstName, lastName, email, profilePic }, context) => {
+    updateUser: async (
+      parent,
+      { firstName, lastName, email, profilePic },
+      context
+    ) => {
       const contextUserId = context.user._id;
       try {
         const updatedUser = await User.findOneAndUpdate(
@@ -179,8 +173,7 @@ const resolvers = {
         console.log("updatedUser", updatedUser);
         console.log("contextUserId", contextUserId);
         return { updatedUser };
-      }
-      catch (error) {
+      } catch (error) {
         console.error("Error updating user:", error);
         throw new Error("Error updating user");
       }
@@ -188,24 +181,20 @@ const resolvers = {
     deleteUser: async (parent, { username }, context) => {
       const contextUserId = context.user._id;
       try {
-        const deletedUser = await User.findOneAndDelete(
-          { _id: contextUserId }
-        );
+        const deletedUser = await User.findOneAndDelete({ _id: contextUserId });
         console.log("deletedUser", deletedUser);
         console.log("contextUserId", contextUserId);
         return { deletedUser };
-      }
-      catch (error) {
+      } catch (error) {
         console.error("Error deleting user:", error);
         throw new Error("Error deleting user");
       }
     },
-    
+
     devDelUser: async (parent, { userId }) => {
       const user = await User.findOneAndDelete({ _id: userId });
-      console.log('user deleted', user);
+      console.log("user deleted", user);
     },
-
 
     login: async (parent, { username, password }) => {
       console.log("hit login");
@@ -223,14 +212,17 @@ const resolvers = {
       return { token, user };
     },
 
-    addPayment: async (_, { userId, chargeId, amount, currency, description }) => {
+    addPayment: async (
+      _,
+      { userId, chargeId, amount, currency, description }
+    ) => {
       try {
         const payment = new Payment({
           userId,
           chargeId,
           amount,
           currency,
-          description
+          description,
         });
         await payment.save();
         return payment;
@@ -238,7 +230,7 @@ const resolvers = {
         throw new Error(err);
       }
     },
-    
+
     uploadFile: async (_, { file }) => {
       try {
         const { createReadStream, filename, mimetype } = await file;
@@ -255,7 +247,7 @@ const resolvers = {
               else resolve(result);
             }
           );
-        fileStream.pipe(cloudStream);
+          fileStream.pipe(cloudStream);
         });
 
         // Return metadata and URL
@@ -268,6 +260,78 @@ const resolvers = {
       } catch (error) {
         throw new Error(`Failed to upload file: ${error}`);
       }
+    },
+
+    upVote: async (parent, { capsuleId, postId }) => {
+      const capsuleIdObject = new ObjectId(capsuleId);
+      const postIdObject = new ObjectId(postId);
+
+      const updatedPost = await Capsule.findOneAndUpdate(
+        { _id: capsuleIdObject, "posts._id": postIdObject },
+        { $inc: { "posts.$.upVotes": 1 } },
+        { new: true }
+      );
+
+      if (!updatedPost) {
+        console.log("post not found");
+        return 0;
+      }
+
+      const results = updatedPost.posts.find((post) =>
+        post._id.equals(postIdObject)
+      );
+      console.log("post found: " + results.upVotes);
+
+      return results.upVotes;
+    },
+
+    downVote: async (parent, { capsuleId, postId }) => {
+      const capsule = await Capsule.findById({ _id: capsuleId });
+
+      const postIdObject = new ObjectId(postId);
+      const post = capsule.posts.find((post) => post._id.equals(postIdObject));
+      if (!post) {
+        console.log("post not found");
+        return null;
+      }
+
+      console.log("post found");
+
+      const upVote = post.upVotes + 1;
+
+      const updatedPost = await Capsule.findOneAndUpdate(
+        { _id: capsuleId, "posts._id": postIdObject },
+        { $set: { "posts.$.downVotes": upVote } },
+        { new: true }
+      );
+
+      return updatedPost;
+    },
+
+    addComment: async (parent, { capsuleId, postId, text, author }) => {
+      const capsule = await Capsule.findById({ _id: capsuleId });
+
+      const postIdObject = new ObjectId(postId);
+      const post = capsule.posts.find((post) => post._id.equals(postIdObject));
+      if (!post) {
+        console.log("post not found");
+        return null;
+      }
+
+      console.log("post found");
+
+      const newComment = {
+        text,
+        author,
+      };
+
+      const updatedPost = await Capsule.findOneAndUpdate(
+        { _id: capsuleId, "posts._id": postIdObject },
+        { $push: { "posts.$.comments": newComment } },
+        { new: true }
+      );
+
+      return updatedPost;
     },
   },
 };
